pyproject.toml
[tool.poetry]
name = "jack-netbridge"
version = "0.2.0"
description = "Send and receive Jack MIDI data as multicast streams over network."
license = "GNU GPL v3"
authors = ["omnitonal.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
JACK-Client = "^0.5.4"
numpy = "^1.26.0"
mido = "^1.3.0"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jack_netbridge = "jack_netbridge.jack_netbridge:main"
